Finish Javascript additions to calculator project
The commit adds in the responsive calculations of the calculator. There
is a bug where if you add in additional numbers after pressing the '='
button it will just add in that number onto the end of the previous
answer. I am not going to fix it because it could be considered a
feature.